// src/common/types/news.ts
export type Article = {
    id: string;
    title: string;
    description: string;
    url: string;
    urlToImage: string;
    publishedAt: string;
    content?: string;
    source: {
        id: string;
        name: string;
    };
    category?: string;
};

export const categories = [
    "all",
    "business",
    "entertainment",
    "general",
    "health",
    "science",
    "sports",
    "technology",
] as const; 